@@ -40,6 +40,22 @@ public class ex17 {
         // incorrect identity value.
         testSum(1L, true);
 
+        // Run the product reduction test sequentially with the
+        // correct identity value.
+        testProd(1L, false);
+
+        // Run the product reduction test in parallel with the
+        // correct identity value.
+        testProd(1L, true);
+
+        // Run the product reduction test sequentially with an
+        // incorrect identity value.
+        testProd(0L, false);
+
+        // Run the product reduction test in parallel with an
+        // incorrect identity value.
+        testSum(0L, true);
+
         // Reduce partial results into a string using a sequential
         // stream and the three parameter version of reduce().
         buggyStreamReduce3(false);
@@ -47,16 +63,29 @@ public class ex17 {
         // Reduce partial results into a string using a parallel
         // stream and the three parameter version of reduce().
         buggyStreamReduce3(true);
+
+        // Reduce partial results into a string using a sequential
+        // stream and string concatenation with reduce().
+        streamReduceConcat(false);
+
+        // Reduce partial results into a string using a parallel
+        // stream and string concatenation with reduce().
+        streamReduceConcat(true);
+
+        // Collect partial results into a string using a sequential
+        // stream together with collect() and StringJoiner.
+        streamCollectJoining(false);
+
+        // Collect partial results into a string using a parallel
+        // stream together with collect() and StringJoiner.
+        streamCollectJoining(true);
     }
 
     /**
      * Print out the results of subtracting the first 100 numbers.
      * If @a parallel is true then a parallel stream is used, else a
-     * sequential stream is used.  When a sequential stream is used
-     * the results of this test will be correct since the subtraction
-     * is done in encounter order.  When a parallel stream is used,
-     * however, the results of this test will be incorrect since
-     * subtraction is not associative.
+     * sequential stream is used.  The results for each of these tests
+     * will differ since subtraction is not associative.
      */
     private static void testDifferenceReduce(boolean parallel) {
         LongStream rangeStream = LongStream
@@ -66,11 +95,11 @@ public class ex17 {
             rangeStream.parallel();
 
         long difference = rangeStream
-            .reduce(0L,
+            .reduce(1L,
                     (x, y) -> x - y);
 
         System.out.println((parallel ? "Parallel" : "Sequential")
-                           + " difference of first 100 numbers = " 
+                           + " difference of first 100 numbers = "
                            + difference);
     }
 
@@ -81,7 +110,7 @@ public class ex17 {
      * sequential stream is used.  When a sequential or parallel
      * stream is used with an identity of 0 the results of this test
      * will be correct.  When a sequential or parallel stream is used
-     * with an identity of 0, however, results of this test will be
+     * with an identity of 1, however, results of this test will be
      * incorrect.
      */
     private static void testSum(long identity,
@@ -105,6 +134,35 @@ public class ex17 {
     }
 
     /**
+     * Print out the results of multiplying the first 100 numbers,
+     * using @a identity as the initial value of the summation.  If @a
+     * parallel is true then a parallel stream is used, else a
+     * sequential stream is used.  When a sequential or parallel
+     * stream is used with an identity of 1 the results of this test
+     * will be correct.  When a sequential or parallel stream is used
+     * with an identity of 0, however, results of this test will be
+     * incorrect.
+     */
+    private static void testProd(long identity,
+                                 boolean parallel) {
+        LongStream rangeStream = LongStream
+            .rangeClosed(1, 10);
+
+        if (parallel)
+            rangeStream.parallel();
+
+        long product = rangeStream
+            .reduce(identity,
+                    (x, y) -> x * y);
+
+        System.out.println((parallel ? "Parallel" : "Sequential")
+                           + " product of first 10 numbers with identity "
+                           + identity
+                           + " = "
+                           + product);
+    }
+
+    /**
      * Reduce partial results into a StringBuilder using the three
      * parameter version of reduce().  If @a parallel is true then a
      * parallel stream is used, else a sequential stream is used.
@@ -118,13 +176,13 @@ public class ex17 {
     private static void buggyStreamReduce3(boolean parallel) {
         System.out.println("\n++Running the "
                            + (parallel ? "parallel" : "sequential")
-                           + "BuggyStreamReduce3 implementation");
+                           + "buggyStreamReduce3 implementation");
 
         List<String> allWords =
-            Arrays.asList("The quick brown fox jumps over the lazy dog\n",
-                          "A man, a plan, a canal: Panama\n",
-                          "Now is the time for all good people " 
-                          + "to come to the aid of their party");
+            List.of("The quick brown fox jumps over the lazy dog\n",
+                    "A man, a plan, a canal: Panama\n",
+                    "Now is the time for all good people " 
+                    + "to come to the aid of their party");
 
         // Record the start time.
         long startTime = System.nanoTime();
@@ -152,6 +210,115 @@ public class ex17 {
                     StringBuilder::append)
             // Create a string.
             .toString();
+
+
+        // Record the stop time.
+        long stopTime = (System.nanoTime() - startTime) / 1_000_000;
+
+        System.out.println("The time to collect "
+                           + words.split("\\s+").length
+                           + " words took "
+                           + stopTime
+                           + " milliseconds.  Here are the words:\n"
+                           + words);
+    }
+
+    /**
+     * Reduce partial results into a String using reduce() with string
+     * concatenation (i.e., the '+' operator).  If @a parallel is
+     * true then a parallel stream is used, else a sequential stream
+     * is used.  This solution is correct, but inefficient due to the
+     * overhead of string concatentation.
+     */
+    private static void streamReduceConcat(boolean parallel) {
+        System.out.println("\n++Running the "
+                           + (parallel ? "parallel" : "sequential")
+                           + "streamReduceConcat implementation");
+
+        List<String> allWords =
+            List.of("The quick brown fox jumps over the lazy dog\n",
+                    "A man, a plan, a canal: Panama\n",
+                    "Now is the time for all good people " 
+                    + "to come to the aid of their party");
+
+        // Record the start time.
+        long startTime = System.nanoTime();
+
+        Stream<String> wordStream = allWords
+            // Convert the list into a stream (which uses a
+            // spliterator internally).
+            .stream();
+
+        if (parallel)
+            // Convert to a parallel stream.
+            wordStream.parallel();
+
+        // A "real" application would likely do something
+        // interesting with the words at this point.
+
+        // Create a string that contains all the strings appended together.
+        String words = wordStream
+            // Use reduce() to append all the strings in the stream.
+            // This implementation works with both sequential and
+            // parallel streams, but it's inefficient since it
+            // requires string concatenation.
+            .reduce(new String(),
+                    (x, y) -> x + y);
+
+
+        // Record the stop time.
+        long stopTime = (System.nanoTime() - startTime) / 1_000_000;
+
+        System.out.println("The time to collect "
+                           + words.split("\\s+").length
+                           + " words took "
+                           + stopTime
+                           + " milliseconds.  Here are the words:\n"
+                           + words);
+    }
+
+    /**
+     * Collect partial results into a string using a parallel stream
+     * together with collect() and joining().  If @a parallel is
+     * true then a parallel stream is used, else a sequential stream
+     * is used.  When a sequential stream or a parallel stream is used
+     * the results of this test will be correct due to the use of a
+     * mutable object (StringJoiner) with collect(), which works
+     * correctly in this case.
+     */
+    private static void streamCollectJoining(boolean parallel) {
+        System.out.println("\n++Running the "
+                           + (parallel ? "parallel" : "sequential")
+                           + "streamCollectJoining implementation");
+
+        List<String> allWords =
+            List.of("The quick brown fox jumps over the lazy dog\n",
+                    "A man, a plan, a canal: Panama\n",
+                    "Now is the time for all good people " 
+                    + "to come to the aid of their party");
+
+        // Record the start time.
+        long startTime = System.nanoTime();
+
+        Stream<String> wordStream = allWords
+            // Convert the list into a stream (which uses a
+            // spliterator internally).
+            .stream();
+
+        if (parallel)
+            // Convert to a parallel stream.
+            wordStream.parallel();
+
+        // A "real" application would likely do something interesting
+        // with the words at this point.
+
+        // Create a string that contains all the strings appended
+        // together.
+        String words = wordStream
+            // Use collect() to append all the strings in the stream.
+            // This implementation works when used with either a
+            // sequential or a parallel stream.
+            .collect(joining());
 
         // Record the stop time.
         long stopTime = (System.nanoTime() - startTime) / 1_000_000;
