@@ -431,7 +431,8 @@ public class MainActivity
             f.delete();
             deletedFiles++;
         }
-        currentFolder.delete();
+        // Don't delete the current folder.
+        // currentFolder.delete();
         return deletedFiles;
     }
 
