@@ -1,6 +1,9 @@
 package livelessons.filters;
 
 import livelessons.utils.Image;
+import livelessons.utils.Options;
+
+import java.io.File;
 
 /**
  * An abstract class that defines an interface for applying filtering
@@ -13,7 +16,7 @@ public abstract class Filter {
     /**
      * Debug logging tag.
      */
-    protected static final String TAG = "Filter";
+    protected final String TAG = Filter.class.getName();
 
     /**
      * The name of the filter, which defaults to the "canonical name"
@@ -27,6 +30,7 @@ public abstract class Filter {
     public Filter() {
         // Default uses the class name without the package prefix.
     	mName = this.getClass().getSimpleName();
+
     }
 
     /**
@@ -67,5 +71,13 @@ public abstract class Filter {
      */
     public String getName() {
         return mName;
+    }
+
+    /**
+     * Gets the file path where a filtered image will be stored.
+     */
+    public File getFilePath() {
+        return new File(Options.instance().getDirectoryPath(),
+                        getName());
     }
 }
