@@ -2,6 +2,8 @@ package livelessons.filters;
 
 import livelessons.utils.Image;
 
+import java.io.File;
+
 /**
  * The NullFilter will return the image as it was downloaded.  It's
  * purpose is to show the original image, as well as to exemplify how
@@ -12,9 +14,16 @@ import livelessons.utils.Image;
 public class NullFilter 
        extends Filter {
     /**
-     * Default constructor is a no-op.
+     * Constructs a default NullFilter and creates the output
+     * directory for this filter.
      */
-    public NullFilter() {}
+    public NullFilter() {
+        // Create the path to the filter directory.
+        File externalFile = getFilePath();
+
+        // Create the filter directory.
+        externalFile.mkdirs();
+    }
     
     /**
      * Constructors for the NullFilter. See GrayScaleFilter for
