@@ -6,6 +6,7 @@ import java.util.Arrays;
 import java.util.List;
 import java.util.concurrent.CompletableFuture;
 import java.util.function.Predicate;
+import java.util.stream.Stream;
 
 /**
  * A utility class containing helpful methods for manipulating various
@@ -28,7 +29,7 @@ public class StreamsUtils {
      *         joined results.
      */
     public static <T> CompletableFuture<List<T>> 
-        joinAll (List<CompletableFuture<T>> futureList) {
+        joinAll(List<CompletableFuture<T>> futureList) {
         // Use CompletableFuture.allOf() to obtain a CompletableFuture
         // that will itself be complete when all CompletableFutures in
         // futureList parameter have completed.
@@ -36,24 +37,86 @@ public class StreamsUtils {
             allDoneFuture = CompletableFuture.allOf
             (futureList.toArray(new CompletableFuture[futureList.size()]));
 
-        // When all futures have completed get a CompletableFuture to
+        // When all futures have completed return a CompletableFuture to
         // a list of joined elements of type T.
-        CompletableFuture<List<T>> allDoneList = allDoneFuture
+        return allDoneFuture
             .thenApply(v -> futureList
                        // Convert futureList into a stream of
                        // completable futures.
                        .stream()
 
-                       // Use map() to join() all completablefutures
+                       // Use map() to join() all completable futures
                        // and yield objects of type T.  Note that
                        // join() should never block.
                        .map(CompletableFuture::join)
 
                        // Collect the results of type T into a list.
                        .collect(toList()));
+    }
 
-        // Return the CompletableFuture. 
-        return allDoneList;
+    /**
+     * Create a CompletableFuture that, when completed, will convert
+     * all the completed CompletableFutures in the @a futureList
+     * parameter into a list of joined results.
+     *
+     * @param futureList A list of completable futures.
+     * @return A CompletableFuture to a stream that will contain all the
+     *         joined results.
+     */
+    public static <T> CompletableFuture<Stream<T>>
+        joinAllStream(List<CompletableFuture<T>> futureList) {
+        // Use CompletableFuture.allOf() to obtain a CompletableFuture
+        // that will itself be complete when all CompletableFutures in
+        // futureList parameter have completed.
+        CompletableFuture<Void>
+            allDoneFuture = CompletableFuture.allOf
+            (futureList.toArray(new CompletableFuture[futureList.size()]));
+
+        // When all futures have completed return a CompletableFuture to
+        // a list of joined elements of type T.
+        return allDoneFuture
+            .thenApply(v -> futureList
+                       // Convert futureList into a stream of
+                       // completable futures.
+                       .stream()
+
+                       // Use map() to join() all completable futures
+                       // and yield objects of type T.  Note that
+                       // join() should never block.
+                       .map(CompletableFuture::join));
+    }
+
+    /**
+     * Create a CompletableFuture that, when completed, will convert
+     * all the completed CompletableFutures in the {@code futureStream}
+     * parameter into a list of joined results.
+     *
+     * @param futureStream A stream of completable futures
+     * @return A CompletableFuture to a stream that will contain all the
+     *         joined results.
+     */
+    public static <T> CompletableFuture<Stream<T>>
+        joinAllStream(Stream<CompletableFuture<T>> futureStream) {
+        // Use CompletableFuture.allOf() to obtain a CompletableFuture
+        // that will itself be complete when all CompletableFutures in
+        // futureStream parameter have completed.
+        CompletableFuture<T>[] futures =
+            futureStream.toArray(CompletableFuture[]::new);
+        CompletableFuture<Void> allDoneFuture =
+                CompletableFuture.allOf(futures);
+
+        // When all futures have completed return a CompletableFuture to
+        // a list of joined elements of type T.
+        return allDoneFuture
+            .thenApply(v -> Arrays
+                       // Convert futureList into a stream of
+                       // completable futures.
+                       .stream(futures)
+
+                       // Use map() to join() all completable futures
+                       // and yield objects of type T.  Note that
+                       // join() should never block.
+                       .map(CompletableFuture::join));
     }
 
     /**
