@@ -1,114 +1,145 @@
 package livelessons.streams;
 
 import livelessons.filters.Filter;
-
-import static java.util.stream.Collectors.toList;
+import livelessons.utils.Image;
 
 import java.net.URL;
 import java.util.Iterator;
 import java.util.List;
+import java.util.Optional;
 import java.util.concurrent.CompletableFuture;
+import java.util.concurrent.ForkJoinPool;
 import java.util.stream.Stream;
 
-import livelessons.utils.Image;
-import livelessons.utils.StreamsUtils;
-import livelessons.filters.FilterDecoratorWithImage;
+import static livelessons.utils.StreamOfFuturesCollector.toFuture;
 
 /**
  * This asynchronous implementation strategy customizes the
  * ImageStreamCompletableFutureBase super class to download, process,
- * and store images asynchronously and concurrently in a thread in the
- * executor's thread pool.
+ * and store images asynchronously in a common fork-join thread pool.
  */
 public class ImageStreamCompletableFuture1
        extends ImageStreamCompletableFutureBase {
-    /**
+     /**
      * Constructor initializes the superclass and data members.
      */
     public ImageStreamCompletableFuture1(Filter[] filters,
                                          Iterator<List<URL>> urlListIterator) {
-        super(filters, urlListIterator);
+        super(filters, 
+              urlListIterator);
     }
 
     /**
-     * Use Java 8 CompletableFutures to download, process, and store
-     * images concurrently and asynchronously.
+     * A hook method that's also a template method.  It assigns the
+     * executor to the common fork-join pool and calls up to the
+     * superclass start the processing.
+     */
+    @Override
+    protected void initiateStream() {
+        // Set the executor to the common fork-join pool.
+        setExecutor(ForkJoinPool.commonPool());
+
+        // Call up to superclass to start the processing.
+        super.initiateStream();
+    }
+
+    /**
+     * This hook method uses the Java completable futures framework to
+     * download, process, and store images in asynchronously.
      */
     @Override
     protected void processStream() {
         // Get the input URLs.
         List<URL> urls = getInput();
 
-        // Create a list of completable futures to filtered images.
-        List<CompletableFuture<Image>> listOfFilteredImageFutures = urls
+        urls
             // Convert the URLs in the input list into a sequential
             // stream.
             .stream()
 
-            // Use filter() to ignore URLs that are already cached locally,
-            // i.e., only download non-cached images.
-            .filter(StreamsUtils.not(this::urlCached))
+            // Use map() to call checkUrlCachedAsync(), which
+            // asynchronously checks if a URL is already cached.
+            .map(this::checkUrlCachedAsync)
 
-            // Use map() to transform each URL to a completable future
-            // to an image (i.e., asynchronously download each image
-            // via its URL).
+            // Use map() to call downloadImageAsync(), which
+            // transforms each non-cached URL to a future to an image
+            // (i.e., asynchronously download each image via its URL).
             .map(this::downloadImageAsync)
 
-            // Use flatMap() to create a stream containing completable
-            // futures to multiple filtered/stored versions of each
-            // image.
+            // Use flatMap() to call applyFiltersAsync(), which
+            // returns a future to a stream of filtered images that
+            // are not already cached locally.
             .flatMap(this::applyFiltersAsync)
 
-            // Terminate the stream and collect the results into
-            // list of completable futures to images.
-            .collect(toList());
+            // Trigger intermediate processing and create a future
+            // that can be used to wait for all operations associated
+            // with the stream of futures to complete.
+            .collect(toFuture())
 
-        // Create a CompletableFuture that can be used to wait for all
-        // operations associated with the futures to complete.
-        CompletableFuture<List<Image>> allImagesDone =
-                StreamsUtils.joinAll(listOfFilteredImageFutures);
+            // thenAccept() is called when all the futures in the
+            // stream complete their asynchronous processing.
+            .thenAccept(stream ->
+                        // Log the results.
+                        log(stream
+                            // Remove any empty optionals.
+                            .flatMap(Optional::stream),
+                            // For JDK 8 you'll need to use
+                            // .filter(Optional::isPresent)
+                            // .map(Optional::get),
+                            urls.size()))
 
-        // Print the results.
-        System.out.println(TAG
-                           + ": processing of "
-                           // This call blocks until all the images
-                           // are downloaded, processed, and stored.
-                           + allImagesDone.join().size()
-                           + " image(s) from "
-                           + urls.size() 
-                           + " urls is complete");
+            // Wait until all the images have been downloaded,
+            // processed, and stored.
+            .join();
     }
 
     /**
-     * Apply filters asynchronously and concurrently to the @a
-     * imageFuture after it finishes downloading and store the results
-     * in output files on the local computer.
+     * Asynchronously download an image from the {@code urlFuture} parameter.
+     *
+     * @param urlFuture A future the URL to download
+     * @return A future that completes when the image finishes downloading
      */
-    private Stream<CompletableFuture<Image>> applyFiltersAsync(CompletableFuture<Image> imageFuture) {
+    private CompletableFuture<Optional<Image>> downloadImageAsync
+              (CompletableFuture<Optional<URL>> urlFuture) {
+        // Return a future that completes when the image finishes
+        // downloading.
+        return urlFuture
+            // Use the executor to asynchronously download an image
+            // when urlFuture completes.
+            .thenApplyAsync(urlOpt ->
+                            urlOpt
+                            // Download non-null URLs.
+                            .map(this::blockingDownload),
+                            // Use the common fork-join pool.
+                            getExecutor());
+    }
+
+    /**
+     * Asynchronously apply filters to the {@code imageFuture} after
+     * it finishes downloading and store the results in output files
+     * on the local computer.
+     *
+     * @param imageFuture A future to an image that's being downloaded
+     @ return A stream of completable futures to images that are being filtered/stored
+    */
+    private Stream<CompletableFuture<Optional<Image>>> applyFiltersAsync
+              (CompletableFuture<Optional<Image>> imageFuture) {
         return mFilters
             // Convert the list of filters to a sequential stream.
             .stream()
 
-            // Use map() to create a completable future to a
-            // FilterDecoratorWithImage object for each filter/image.
-            .map(filter ->
-                 // Returns a new CompletionStage that, when this
-                 // stage completes normally, is executed with this
-                 // stage's result as the argument to the supplied
-                 // lambda expression.
-                 imageFuture.thenApply(image ->
-                                       makeFilterDecoratorWithImage(filter,
-                                                                    image)))
-                                                 
-            // Asynchronously filter the image and store it in an
-            // output file.
-            .map(filterFuture ->
-                 // Returns a new CompletionStage that, when this
-                 // stage completes normally, is executed with this
-                 // stage's result as the argument to the supplied
-                 // lambda expression.
-                 filterFuture.thenCompose(filter ->
-                                          CompletableFuture.supplyAsync(filter::run,
-                                                                        getExecutor())));
+            // Use map() to filter each image asynchronously.
+            .map(filter -> imageFuture
+                 // Asynchronously apply a filter action after the
+                 // previous stage completes.
+                 .thenApplyAsync(imageOpt ->
+                                 imageOpt
+                                 // Create and apply the filter to the
+                                 // image.
+                                 .map(image ->
+                                      makeFilterDecoratorWithImage(filter,
+                                                                   image).run()),
+                                 // Run in the common fork-join pool.
+                                 getExecutor()));
     }
 }
