@@ -1,142 +1,184 @@
 package livelessons.streams;
 
 import livelessons.filters.Filter;
-import java.net.URL;
-import static java.util.stream.Collectors.toList;
-
-import java.util.Iterator;
-import java.util.List;
-import java.util.concurrent.CompletableFuture;
-
 import livelessons.utils.Image;
 import livelessons.utils.StreamsUtils;
-import livelessons.filters.FilterDecoratorWithImage;
-import static java.util.stream.Collectors.summingInt;
+
+import java.net.URL;
+import java.util.Iterator;
+import java.util.List;
+import java.util.Optional;
+import java.util.concurrent.CompletableFuture;
+import java.util.concurrent.Executors;
+import java.util.function.Function;
+import java.util.stream.Stream;
+
+import static java.util.stream.Collectors.toList;
+import static livelessons.utils.StreamOfFuturesCollector.toFuture;
 
 /**
  * This is another asynchronous implementation strategy that
  * customizes the ImageStreamCompletableFutureBase super class to
- * download, process, and store images asynchronously and concurrently
- * in a thread in the executor's thread pool.
+ * download, process, and store images asynchronously in the
+ * designated thread pool.
  */
 public class ImageStreamCompletableFuture2
        extends ImageStreamCompletableFutureBase {
     /**
+     * Maximum number of threads in a fixed-size thread pool.
+     */
+    private final int sMAX_THREADS = 100;
+
+    /**
      * Constructor initializes the superclass and data members.
      */
     public ImageStreamCompletableFuture2(Filter[] filters,
-                                        Iterator<List<URL>> urlListIterator) {
-        super(filters, urlListIterator);
+                                         Iterator<List<URL>> urlListIterator) {
+        super(filters, 
+              urlListIterator);
     }
 
     /**
-     * Use Java 8 CompletableFutures to download, process, and store
-     * images concurrently.
+     * A hook method that's also a template method.  It sets the
+     * executor to a fixed-size thread pool and calls up to the
+     * superclass start the processing.
+     */
+    @Override
+    protected void initiateStream() {
+        // The thread pool size is the smaller of (1) the number of
+        // filters times the number of images to download and (2)
+        // sMAX_THREADS (which prevents allocating excessive threads).
+        int threadPoolSize = Math.min(mFilters.size() * getInput().size(),
+                                      sMAX_THREADS);
+
+        // Initialize the Executor with appropriate pool of threads.
+        setExecutor(Executors.newFixedThreadPool(threadPoolSize));
+
+        // Call up to superclass to start the processing.
+        super.initiateStream();
+    }
+
+    /**
+     * This hook method uses the Java completable future framework to
+     * download, process, and store images in asynchronously.
      */
     @Override
     protected void processStream() {
         // Get the input URLs.
         List<URL> urls = getInput();
 
-        // Create a list of completable futures to a list of images.
-        final List<CompletableFuture<List<Image>>> listOfFilteredImageFutures = urls
+        // A future to a stream of URLs.
+        Stream<CompletableFuture<Optional<URL>>> urlStream = urls
             // Convert the URLs in the input list into a sequential
             // stream.
             .stream()
 
-            // Use filter() to ignore URLs that are already cached locally,
-            // i.e., only download non-cached images.
-            .filter(StreamsUtils.not(this::urlCached))
+            // Use map() to ignore URLs that are already cached
+            // locally, i.e., only download non-cached images.
+            .map(this::checkUrlCachedAsync);
 
-            // Use map() to transform each URL to a completable future
-            // to an image (i.e., asynchronously download each image
-            // via its URL).
-            .map(this::downloadImageAsync)
+        StreamsUtils
+            // Create a future that is used to wait for
+            // all futures in urlStream to complete.
+            .joinAllStream(urlStream)
 
-            // Use map() to call makeFilterDecorators(), which returns
-            // a List of FilterDecoratorWithImage objects associated
-            // with a completable future.
-            .map(this::makeFilterDecorators)
+            // thenAccept() is called when all the futures in the
+            // stream complete their processing.
+            .thenAccept(stream -> {
+                    // Create a stream of completable futures to
+                    // filtered images.
+                    Stream<CompletableFuture<Stream<Image>>> futureStream = stream
+                        // Remove all cached URLs.
+                        .flatMap(Optional::stream)
+                        // For JDK 8 you'll need to use 
+                        // .filter(Optional::isPresent)
+                        // .map(Optional::get)
 
-            // Use map() to call the applyFiltersAsync() method
-            // reference, which returns a list of filtered Image
-            // futures.
-            .map(this::applyFiltersAsync)
+                        // Use map() to call downloadImageAsync(),
+                        // which transforms each URL to a completable
+                        // future to an image (i.e., asynchronously
+                        // download each image via its URL).
+                        .map(this::downloadImageAsync)
+
+                        // Use map() to call applyFiltersAsync(),
+                        // which creates a stream of completable
+                        // futures to multiple filtered/stored
+                        // versions of each image.
+                        .map(this::applyFiltersAsync);
+
+                    StreamsUtils
+                        // Create a future that is used to wait for
+                        // all futures in futureStream to complete.
+                        .joinAllStream(futureStream)
+
+                        // thenAccept() is called when all the futures
+                        // in the stream complete their processing.
+                        .thenAccept(resultsStream ->
+                                    // Flatten the stream of streams
+                                    // and log the results.
+                                    log(resultsStream
+                                        .flatMap(Function.identity()),
+                                                 urls.size()))
+                        
+                        // Wait until all images have been downloaded,
+                        // processed, and stored.
+                        .join();
+                })
             
-            // Terminate the stream, which returns a list of
-            // completable futures to a list of images.
-            .collect(toList());
-
-        // Create a CompletableFuture that can be used to wait for all
-        // operations associated with the futures to complete.
-        CompletableFuture<List<List<Image>>> allImagesDone =
-                StreamsUtils.joinAll(listOfFilteredImageFutures);
-
-        // The call to join() is needed here to blocks the calling
-        // thread until all the futures have completed.
-        Integer imagesProcessed = allImagesDone
-            // join() returns a list of list of images, so we convert
-            // this to a stream and then sum up the size of each of
-            // these lists to get the final count of images processed.
-            .join()
-            .stream()
-            .collect(summingInt(List::size));
-
-        System.out.println(TAG
-                           + ": processing of "
-                           + imagesProcessed
-                           + " image(s) from "
-                           + urls.size() 
-                           + " urls is complete");
+            // Wait until all images have been downloaded, processed,
+            // and stored.
+            .join();
     }
 
     /**
-     * A factory method that makes all the image processing filters
-     * for the @a imageFuture after its downloaded.
+     * Asynchronously download an image from the {@code url} parameter.
+     *
+     * @param url The URL to download
+     * @return A future that completes when the image finishes downloading
      */
-    private CompletableFuture<List<FilterDecoratorWithImage>> makeFilterDecorators
-                (CompletableFuture<Image> imageFuture) {
-        // Returns a new CompletionStage that, when this stage
-        // completes normally, is executed with this stage's result as
-        // the argument to the supplied lambda expression.
-        return imageFuture.thenApply(image -> mFilters
-            // Convert all the image filters into a sequential stream.
-            .stream()
-
-            // Create a FilterDecoratorWithImage object for each
-            // filter/image combo.
-            .map(filter -> makeFilterDecoratorWithImage(filter, image))
-
-            // Return a completable future to a list of
-            // FilterDecoratorWithImage objects.
-            .collect(toList()));
+    private CompletableFuture<Image> downloadImageAsync(URL url) {
+        // Return a future that completes when the image finishes
+        // downloading.
+        return CompletableFuture
+            // Use the executor to asynchronously download an image
+            // when urlFuture completes.
+            .supplyAsync(() -> downloadImage(url),
+                         getExecutor());
     }
 
     /**
      * Asynchronously apply all the filters to images and store them
      * in an output file on the local computer.
-     */
-    private CompletableFuture<List<Image>> applyFiltersAsync
-      (CompletableFuture<List<FilterDecoratorWithImage>> decoratedFiltersWithImageFuture) {
-        // Returns a new CompletionStage that, when this stage
-        // completes normally, is executed with this stage's result as
-        // the argument to the supplied lambda expression.
-        return decoratedFiltersWithImageFuture.thenCompose(list -> {
-                List<CompletableFuture<Image>> listOfFutures = list
-                    // Converts the list of FilterDecoratorWithImage
-                    // objects into a sequential stream.
-                    .stream()
+     *
+     * @param imageFuture A future to an image that's being downloaded
+     @ return A completable future to a stream of images that are being filtered/stored 
+    */
+    private CompletableFuture<Stream<Image>> applyFiltersAsync
+        (CompletableFuture<Image> imageFuture) {
+        // Return a completable future to a stream of images that are
+        // being processed.
+        return imageFuture
+            // thenCompose() is called after the image download
+            // completes.  It works like flatMap(), i.e., it returns a
+            // future to a stream of futures.
+            .thenCompose(image -> mFilters
+                         // Convert image filters to a sequential
+                         // stream.
+                         .stream()
 
-                    // Asynchronously apply a filter to an Image.
-                    .map(this::filterImageAsync)
+                         // Use map() to asynchronously apply a filter.
+                         .map(filter -> CompletableFuture
+                              // Asynchronously apply filter action.
+                              .supplyAsync(() -> 
+                                           // Create and apply the
+                                           // filter to the image.
+                                           makeFilterDecoratorWithImage
+                                           (filter, image).run(),
+                                           // Run in designated
+                                           // executor.
+                                           getExecutor()))
 
-                    // Collect the list of futures.
-                    .collect(toList());
-
-                // Return a CompletableFuture to a list of images that
-                // will be available when all the CompletableFutures
-                // in listOfFutures complete.
-                return StreamsUtils.joinAll(listOfFutures);
-            });
+                         // Collect a stream of futures.
+                         .collect(toFuture()));
     }
 }
