@@ -14,7 +14,7 @@ import livelessons.filters.FilterDecoratorWithImage;
 
 /**
  * This implementation strategy customizes ImageStreamGang to use a
- * Java 8 stream to download, process, and store images sequentially.
+ * Java stream to download, process, and store images sequentially.
  */
 public class ImageStreamSequential 
        extends ImageStreamGang {
@@ -27,7 +27,7 @@ public class ImageStreamSequential
     }
 
     /**
-     * This hook method uses a Java 8 stream to download, process, and
+     * This hook method uses a Java stream to download, process, and
      * store images sequentially.
      */
     @Override
@@ -73,11 +73,10 @@ public class ImageStreamSequential
             // apply each one to the image.
             .stream()
 
-            // Use map() to create an OutputFilterDecorator for each image.
-            .map(filter -> makeFilterDecoratorWithImage(filter, image))
-
-            // Use map() to apply the OutputFilterDecorator to each
-            // image and store it in an output file.
-            .map(FilterDecoratorWithImage::run);
+            // Use map() to create an OutputFilterDecorator for each
+            // image and run it to filter each image and store it in an
+            // output file.
+            .map(filter -> 
+                 makeFilterDecoratorWithImage(filter, image).run());
     }
 }
