@@ -2,6 +2,7 @@ package livelessons.streamgangs;
 
 import livelessons.utils.Options;
 import livelessons.utils.PhraseMatchSpliterator;
+import livelessons.utils.RunTimer;
 import livelessons.utils.SearchResults;
 
 import java.util.*;
@@ -82,14 +83,10 @@ public class SearchStreamGang
         // Create a new barrier for this iteration cycle.
         mExitBarrier = new CountDownLatch(1);
 
-        // Start timing the test run.
-        startTiming();
-
-        // Start the stream processing.
-        List<List<SearchResults>> results = processStream();
-        
-        // Stop timing the test run.
-        stopTiming();
+        // Execute the test and time how long it runs.
+        List<List<SearchResults>> results =
+            RunTimer.timeRun(this::processStream,
+                             TAG);
 
         // Print the results.
         printResults(TAG, results);
@@ -163,15 +160,15 @@ public class SearchStreamGang
 
         // Print out the titles.
         if (Options.getInstance().isVerbose())
-            printTitles(listOfListOfSearchResults);
+            printPhrases(listOfListOfSearchResults);
     }
 
     /**
-     * Print out the quotes associated with each play title.
+     * Displays the phrases associated with each play.
      */
-    private void printTitles(List<List<SearchResults>> listOfListOfSearchResults) {
-        // Create a map that associates phrases found in the input with
-        // the indices where they were found.
+    private void printPhrases(List<List<SearchResults>> listOfListOfSearchResults) {
+        // Create a map that associates phrases found in the input
+        // with the titles where they were found.
         Map<String, List<SearchResults>> resultsMap = listOfListOfSearchResults
             // Convert the list of lists into a stream of lists.
             .stream()
@@ -179,12 +176,15 @@ public class SearchStreamGang
             // Flatten the lists into a stream of SearchResults.
             .flatMap(List::stream)
 
-            // Collect the SearchResults into a Map by their titles.
-            .collect(groupingBy(SearchResults::getTitle));
+            // Collect the SearchResults into a TreeMap, which will
+            // sort the keys by their titles.
+            .collect(groupingBy(SearchResults::getTitle,
+                                TreeMap::new,
+                                toList()));
 
-        // Print out the results in the map, where each title is
-        // first printed followed by a list of the indices where the
-        // phrase appeared in the input.
+        // Print out the results in the map, where each title is first
+        // printed followed by a list of the indices where the phrase
+        // appeared in the input.
         resultsMap.forEach((key, value)
                      -> {
                          System.out.println("Title \""
@@ -257,36 +257,5 @@ public class SearchStreamGang
         return null; 
     }
 
-    /**
-     * Keeps track of how long the test has run.
-     */
-    private long mStartTime;
-
-    /**
-     * Keeps track of all the execution times.
-     */
-    private List<Long> mExecutionTimes = new ArrayList<>();
-
-    /**
-     * Start timing the test run.
-     */
-    private void startTiming() {
-        // Note the start time.
-        mStartTime = System.nanoTime();
-    }
-
-    /**
-     * Stop timing the test run.
-     */
-    private void stopTiming() {
-        mExecutionTimes.add((System.nanoTime() - mStartTime) / 1_000_000);
-    }
-
-    /**
-     * Return the time needed to execute the test.
-     */
-    public List<Long> executionTimes() {
-        return mExecutionTimes;
-    }
 }
 
