@@ -23,8 +23,8 @@ public class TestDataFactory {
     }
 
     /**
-     * Return the input data in the given @a filename as a list of
-     * Strings.
+     * Return the input data in the given {@code filename} as a list
+     * of CharSequence objects.
      */
     public static List<CharSequence> getInput(String filename,
                                               String splitter) {
@@ -47,7 +47,7 @@ public class TestDataFactory {
                 // Filter out any empty strings.
                 .filter(((Predicate<String>) String::isEmpty).negate())
                 
-                // Collect the results into a string.
+                // Collect results into a list of strings.
                 .collect(toList());
         } catch (Exception e) {
             e.printStackTrace();
@@ -56,8 +56,8 @@ public class TestDataFactory {
     }
 
     /**
-     * Return the input data in the given @a filename as a list of
-     * SharedStrings.
+     * Return the input data in the given {@code filename} as a list of
+     * CharSequence objects.
      */
     public static List<CharSequence> getSharedInput(String filename,
                                                     String splitter) {
@@ -74,20 +74,20 @@ public class TestDataFactory {
                 // the file into a list of Strings.
                 Pattern.compile(splitter)
 
-                       // Creates a stream from the given input
-                       // sequence around matches of this pattern.
+                // Creates a stream from the given input sequence
+                // around matches of this pattern.
                 .splitAsStream(bytes)
 
-                       // Filter out any empty strings.
-                       .filter(((Predicate<String>) String::isEmpty).negate())
+                // Filter out any empty strings.
+                .filter(((Predicate<String>) String::isEmpty).negate())
 
-                       // Map each String to a SharedString to
-                       // eliminate copying overhead.
-                       .map(string -> 
-                            new SharedString(string.toCharArray()))
+                // Map each string to a SharedString to eliminate
+                // copying overhead.
+                .map(string -> 
+                     new SharedString(string.toCharArray()))
 
-                       // Collect the results into a string.
-                       .collect(toList());
+                // Collect results into a list of char sequences.
+                .collect(toList());
         } catch (Exception e) {
             e.printStackTrace();
             return null;
@@ -95,7 +95,7 @@ public class TestDataFactory {
     }
 
     /**
-     * Return the phrase list in the @a filename as a list of
+     * Return the phrase list in the {@code filename} as a list of
      * non-empty strings.
      */
     public static List<String> getPhraseList(String filename) {
@@ -103,7 +103,7 @@ public class TestDataFactory {
             return Files
                 // Read all lines from filename.
                 .lines(Paths.get(ClassLoader.getSystemResource
-                                        (filename).toURI()))
+                                 (filename).toURI()))
                 // Filter out any empty strings.
                 .filter(((Predicate<String>) String::isEmpty).negate())
 
